fix(template): ship bundle with empty `tns-java-classes.js` chunk
This is required for compatibility reasons (NativeScript 2.5).
When a non-webpacked NativeScript application is deployed to phone, the apk
contains a `tns-java-classes.js` file. If that file is present, it is
dynamically loaded by the Android runtime at startup.
Webpacked NativeScript applications didn't contain that file as it is
not needed. But when the application is deployed to phone, the previous
`tns-java-classes.js` file is not removed from the app folder on the
phone. That causes the Android runtime to try load the file and fail.
Since this commit, all webpacked NativeScript apps will be shipped with
an empty `tns-java-classes.js` file which is harmless and is used to
override the existing file left from previous deployment.

related to https://github.com/NativeScript/NativeScript/issues/3613, https://github.com/NativeScript/NativeScript/issues/3947