fix(template): disable minification of css by css-loader (#154)
When uglify is used, it triggers the minification option of the
css-loader. The css-loader uses internally cssnano to optimize the size
of the stylesheets. Since NativeScript supports only a subset of the CSS language, many of these optimizations may be damaging for the user's css.
Such is the case with `background-position`: center, which gets
translated to `background-position`: 50%, which is currently not working
in NativeScript 3.0 (https://github.com/NativeScript/NativeScript/issues/4201).

closes #135