refactor: Refactor moduleId: __filename to moduleId: module.id, implement compat loader for webpack
The angular 5 framework will typecheck the moduleId is a string,
while in webpack scenarios the moduleId is ignored and relative paths
are handled by the angular compiler and the webpack loaders,
module.id is number and fails the typeschecks.

Implemented a loader that will strip the moduleId: module.id when
compiled with webpack, the moduleId: module.id is still necessary for development without webpack.

As a reminder if we manage to setup webpack for watch and debug,
we can recommend removing all moduleId: module.id occurences and delete the loader.