fix(compiler): reject promise with real error (#350)
* fix(compiler): reject promise with real error

When the webpack compiler fails, the error is not used in the rejection, so the promise is always rejected with `undefined`.
Fix this by respecting the error. Construct real error object as well.

* Force rebuild of application when webpack is used

In case we want to use webpack, force the build by setting nativeChanged property of changesInfo to true.
This is TEMP solution until we have a better one.