fix: Set INIT_CWD env in all before-* hooks
In a long living process, where the project can be changed, we need to set
the INIT_CWD environment variable in all before-* hooks. Currently it is
set only in the `before-watchPatterns` hook. This way, when we use webpack
for a project and later we try to use webpack for another project, we try
to execute some operations based on the INIT_CWD directory, which still
points to our old project. In order to fix this unexpected behavior, set
the INIT_CWD in all before-* hooks, as some of them may be skipped based
on the state of the project.