fix(hooks): `after-watch` hook is not executed
The `after-watch` hook is not executed as it is not registered in the package.json. So it is never created in the projects, so the webpack process is not stopped when LiveSync should stop.
In CLI process this is not a problem, but this is a huge issue for Sidekick, as the webpack watcher keeps watching the project and executes some actions.