refactor: rename tsconfig.esm.json to tsconfig.tns.json (#635)
Currently the `tsconfig.esm.json` file is added to code-sharing projects, but that's
not necessary, because there's already a `tsconfig.tns.json` file
targetting `es2015`. Also, because `tsconfig.esm.json` is the default
config in the webpack config, the webpack config should also be modified
by the `ng add` schematic.

This change will simplify `ng add` and will make it possible to use the
standard webpack config without modifying it through schematics.