fix: project's package.json indentation is not persisted
In case project's package.json uses tabs or multiple spaces for indentation, the postinstall script of nativescript-dev-webpack overwrites it with two spaces whenever it is executed.
The problem is that the plugin tries to modify the package.json and persists it on every operation.
Fix the behavior by checking the indentation character and use it when stringifying the content.
Also compare the current content of the file with the one we will write and skip the write operation in case they match.
Add unit tests for the new methods.