chore: entryPath check should work on Windows
In a previous commit we've changed the check when to execute the `bundle-config-loader` to be based on full path. However, the created RegExp does not match anything on Windows as the paths contain `\`, which are used for escape symbol inside RegExp.
Escape the `\` and update all webpack configs with the new RegExp.