chore: fix liveSync of html files in Angular apps
In case you change `.html` file in Angular application, during `tns run <platform> --bundle`, the application will be restarted and an error will be shown.
This happens as in some cases, the `getCompiledFile` method from Angular compiler does not throw an error when it is unable to find a file, but instead it returns an object with empty outputText and array of errors (errorDependencies).
Fix the liveSync in these cases by checking the result of the Angular compiler and fallback to non-platform specific file in the mentioned case.