fix: stop generating .d.ts on TS compilation
In case `declaration` is set to true in `tsconfig.json` file, the webpack watcher goes in indefinite loop as each change of .ts file leads to generation of new .d.ts files which are also detected by webpack.
To prevent this, ensure declaration is set to false in all compilation cases.