feat: introduce webpack only workflow (#882)
* fix: don't provide fake paths to the {N} CLI - relative to the `app` folder

* feat: remove not needed hooks

* feat: remove webpack compiler logic

* fix: don't copy app_resources to the platforms folder

* fix: respect --env.verbose

* feat: respect production mode based on release option

Implements: https://github.com/NativeScript/nativescript-dev-webpack/issues/911

* fix: watch platform specific files from node_modules

Rel to: https://github.com/NativeScript/nativescript-cli/issues/4480

* fix: don't emit absolute webpack's runtime files

Previously we needed to emit files with full paths as {N} CLI relies on this and expected them in a such format. With the changes for "webpack-only" mode, {N} CLI expects only relative paths. So we need to fix this in order to ensure that runtime.js file will not be transferred on device on change in hmr mode.

* fix: don't process runtime.js files

We needed a special processing for `runtime.js` files as we excluded them when transferring the files on device. As the CLI is filtering and emit only hot-update files we don't need this logic anymore.

* fix: emit runtime files and entry point files

* fix: update webpack config files of demo apps

* fix: don't use short imports in demo apps

* fix: update dependencies of demo apps so they are compatible with 6.0 release