fix(hmr): check for hot update should not create new file
Currently the check for hot update creates a new file in case it does not exist (as the method from tns-core-modules is doing this). This is a problem when trying to install `.ipa` on device and the `.ipa` file contains JavaScript files with HMR enabled.
This may happen in case you run `tns run ios` on device and after command finishes the execution open the project in Xcode and deploy the app from there or uninstall it from device and install the produced `.ipa` manually. The problem in the mentioned scenarios is that the JavaScript file cannot write files in the directory where the `.ipa` is installed.
When `tns run ios` is executed, it livesyncs the files in a different location, so the HMR can create the files there.

To fix the issue check if the hmr file exist before reading its content.