fix: stop generating snapshot for ia64 for older runtimes
When the runtime version is below 6.0.2 we should generate snapshot only for armv7, arm64 and ia32 archs. However, as in the validation schema we have default value for the targetArchs, the logic that should determine if ia64 should be removed from the targetArchs decides the archs are passed by the user and does not strip anything from them. Remove the default values from the JSON Schema - they are calculated in the code, so there's no need to have them on two places. As we can not have our conditional logic in the JSON schema, keep calculation of the default values only in the code.