fix: stop on compilation error in typescript applications
We have a limitation that webpack compilation doesn't stop on error in typescript applications. That was due to the issue in fork-ts-checker-webpack-plugin. After merging [the PR that fixes the issue](https://github.com/TypeStrong/fork-ts-checker-webpack-plugin/pull/337) and releasing 2.0.0 version of the plugin, we can update it on our side and that way webpack compilation will stop on syntax/semantic errors within the application.

Rel to: https://github.com/NativeScript/nativescript-cli/issues/3785