fix: bundle emitted on save without changes
When there are no changes (timestamps are not modified) the chunk is taken from the cache. The cached asset object has the property existsAt set and this will prevent the emit of that chunk - https://github.com/webpack/webpack/blob/4056506488c1e071dfc9a0127daa61bf531170bf/lib/Compiler.js#L326. However when we replace the whole asset, the flag is no longer set and the chunk is emmited every time.