fix: add import of `.css` file into another `.css` file
After changing `css-loader` with `css2json-loader`, we have an issue that module imports of local css files doesn't work anymore - https://github.com/NativeScript/nativescript-dev-webpack/issues/1098. As it turns out that we don't have applications that test this scenario, we decided to add such an import in demo JavaScript application.